Split Android cloud remote packages
