+++ apps/android/data/local/src/androidTest/java/com/flashcardsopensourceapp/data/local/cloud/sync/SyncLocalStoreReviewHistoryEventContractTest.kt
@@ -1,7 +1,7 @@
 package com.flashcardsopensourceapp.data.local.cloud.sync
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.flashcardsopensourceapp.data.local.cloud.remote.RemoteReviewHistoryEvent
+import com.flashcardsopensourceapp.data.local.cloud.remote.sync.RemoteReviewHistoryEvent
 import com.flashcardsopensourceapp.data.local.database.core.AppDatabase
 import com.flashcardsopensourceapp.data.local.database.entities.ReviewLogEntity
 import com.flashcardsopensourceapp.data.local.model.cloud.CloudWorkspaceSummary
