+++ apps/android/data/local/src/androidTest/java/com/flashcardsopensourceapp/data/local/repository/cloudsync/sync/LocalSyncRepositorySyncBlockingTest.kt
@@ -3,7 +3,7 @@ package com.flashcardsopensourceapp.data.local.repository.cloudsync.sync
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.flashcardsopensourceapp.data.local.cloud.remote.CloudRemoteException
 import com.flashcardsopensourceapp.data.local.cloud.remote.CloudRemoteGateway
-import com.flashcardsopensourceapp.data.local.cloud.remote.RemotePushResponse
+import com.flashcardsopensourceapp.data.local.cloud.remote.sync.RemotePushResponse
 import com.flashcardsopensourceapp.data.local.cloud.identity.syncWorkspaceForkRequiredErrorCode
 import com.flashcardsopensourceapp.data.local.database.entities.OutboxEntryEntity
 import com.flashcardsopensourceapp.data.local.database.entities.SyncStateEntity
