+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/cloud/remote/guest/CloudGuestUpgradeRemoteApi.kt
@@ -1,5 +1,7 @@
-package com.flashcardsopensourceapp.data.local.cloud.remote
+package com.flashcardsopensourceapp.data.local.cloud.remote.guest
 
+import com.flashcardsopensourceapp.data.local.cloud.remote.transport.CloudJsonHttpClient
+import com.flashcardsopensourceapp.data.local.cloud.remote.workspace.parseCloudWorkspace
 import com.flashcardsopensourceapp.data.local.cloud.wire.CloudContractMismatchException
 import com.flashcardsopensourceapp.data.local.cloud.wire.optCloudArrayOrNull
 import com.flashcardsopensourceapp.data.local.cloud.wire.optCloudObjectOrNull
