+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/cloud/sync/SyncLocalStore.kt
@@ -2,9 +2,9 @@ package com.flashcardsopensourceapp.data.local.cloud.sync
 
 import com.flashcardsopensourceapp.data.local.cloud.CloudPreferencesStore
 import com.flashcardsopensourceapp.data.local.cloud.identity.WorkspaceIdentityLocalStore
-import com.flashcardsopensourceapp.data.local.cloud.remote.RemoteBootstrapEntry
-import com.flashcardsopensourceapp.data.local.cloud.remote.RemoteReviewHistoryEvent
-import com.flashcardsopensourceapp.data.local.cloud.remote.RemoteSyncChange
+import com.flashcardsopensourceapp.data.local.cloud.remote.sync.RemoteBootstrapEntry
+import com.flashcardsopensourceapp.data.local.cloud.remote.sync.RemoteReviewHistoryEvent
+import com.flashcardsopensourceapp.data.local.cloud.remote.sync.RemoteSyncChange
 import com.flashcardsopensourceapp.data.local.database.core.AppDatabase
 import com.flashcardsopensourceapp.data.local.database.entities.CardEntity
 import com.flashcardsopensourceapp.data.local.database.entities.DeckEntity
